+++ react-components/src/index.ts
@@ -1,9 +1,6 @@
 import './style/normalize.css';
 import './utils/dayjsLocale';
 
-import { ICellEvent } from './components/Table/model/i-cell-event';
-import { TableColumn } from './components/Table/model/public-api';
-
 export * from './components/Button';
 export * from './components/Button/model/public-api';
 
@@ -68,3 +65,5 @@ export * from './components/Card';
 
 export * from './components/Chart';
 export * from './components/Chart/model';
+
+export * from './components/Drawer';